refactor(resource-usage): moved to /v1/resource-usage
